+++ apps/feature-flags/pages/api/v1/flags/[slug].ts
@@ -2,11 +2,17 @@ import type { NextApiRequest, NextApiResponse } from "next";
 // Prisma
 import { getFlagBySlug } from "libs/prisma/utils/flag";
 // Utils
-import { methodNotAllowed, somethingWentWrong } from "utils/helpers/responses";
+import Cors from "cors";
+import { withCors, withHeaders } from "@basestack/utils";
 // Middlewares
 import withRateLimit from "utils/middleware/rateLimit";
-import withHeaders from "utils/middleware/headers";
-import withCors from "utils/middleware/cors";
+
+// More at https://github.com/expressjs/cors
+const cors = Cors({
+  methods: ["GET"],
+  origin: process.env.API_ACCESS_CONTROL_ALLOW_ORIGIN ?? "*",
+  optionsSuccessStatus: 200,
+});
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   switch (req.method) {
@@ -33,17 +39,22 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
       } catch (error: any) {
         return res.status(error.code ?? 400).json({
           error: true,
-          message: error.message ?? somethingWentWrong,
+          message: error.message ?? "Something went wrong",
         });
       }
       break;
 
     default:
       res.setHeader("Allow", ["GET"]);
-      res.status(405).json(methodNotAllowed);
+      res.status(405).json({
+        code: 405,
+        error: true,
+        message: `The HTTP method is not supported at this route.`,
+      });
   }
 };
 
 export default withCors(
+  Cors,
   withRateLimit(withHeaders(["x-project-key", "x-environment-key"], handler)),
 );
